Decode shader codes added to ShaderChunk; Added more geometry examples; Code optimizattions & improvements
