+++ examples/jsm/utils/GeometryCompressionUtils.js
@@ -58,6 +58,7 @@ var GeometryCompressionUtils = {
 		let result;
 		if (encodeMethod == "DEFAULT") {
 
+			// TODO: Add 1 byte to the result, making the encoded length to be 4 bytes. 
 			result = new Uint8Array(count * 3);
 
 			for (let idx = 0; idx < array.length; idx += 3) {
@@ -695,6 +696,7 @@ function PackedPhongMaterial(parameters) {
 
 	]);
 
+	// use some of the shader src from ShaderChunk. 
 	this.vertexShader = [
 		"#define PHONG",
 
@@ -717,91 +719,22 @@ function PackedPhongMaterial(parameters) {
 		ShaderChunk.logdepthbuf_pars_vertex,
 		ShaderChunk.clipping_planes_pars_vertex,
 
-		`#ifdef USE_PACKED_NORMAL
-			#if USE_PACKED_NORMAL == 0
-				vec3 decodeNormal(vec3 packedNormal)
-				{
-					float x = packedNormal.x * 2.0 - 1.0;
-					float y = packedNormal.y * 2.0 - 1.0;
-					vec2 scth = vec2(sin(x * PI), cos(x * PI));
-					vec2 scphi = vec2(sqrt(1.0 - y * y), y);
-					return normalize( vec3(scth.y * scphi.x, scth.x * scphi.x, scphi.y) );
-				}
-			#endif
-
-			#if USE_PACKED_NORMAL == 1
-				vec3 decodeNormal(vec3 packedNormal)
-				{
-					vec3 v = vec3(packedNormal.xy, 1.0 - abs(packedNormal.x) - abs(packedNormal.y));
-					if (v.z < 0.0)
-					{
-						v.xy = (1.0 - abs(v.yx)) * vec2((v.x >= 0.0) ? +1.0 : -1.0, (v.y >= 0.0) ? +1.0 : -1.0);
-					}
-					return normalize(v);
-				}
-			#endif
-
-			#if USE_PACKED_NORMAL == 2
-				vec3 decodeNormal(vec3 packedNormal)
-				{
-					vec3 v = (packedNormal * 2.0) - 1.0;
-					return normalize(v);
-				}
-			#endif
-		#endif`,
-
-		`#ifdef USE_PACKED_POSITION
-			#if USE_PACKED_POSITION == 0
-				uniform mat4 quantizeMatPos;
-			#endif
-		#endif`,
-
-		`#ifdef USE_PACKED_UV
-			#if USE_PACKED_UV == 1
-				uniform mat3 quantizeMatUV;
-			#endif
-		#endif`,
-
-		`#ifdef USE_PACKED_UV
-			#if USE_PACKED_UV == 0
-				vec2 decodeUV(vec2 packedUV)
-				{
-					vec2 uv = (packedUV * 2.0) - 1.0;
-					return uv;
-				}
-			#endif
-
-			#if USE_PACKED_UV == 1
-				vec2 decodeUV(vec2 packedUV)
-				{
-					vec2 uv = ( vec3(packedUV, 1.0) * quantizeMatUV ).xy;
-					return uv;
-				}
-			#endif
-		#endif`,
+		// Shader src: `geometry_decode_pars_vertex`
+		ShaderChunk.geometry_decode_pars_vertex, 
 
 		"void main() {",
 
 		ShaderChunk.uv_vertex,
-
-		`#ifdef USE_UV
-			#ifdef USE_PACKED_UV
-				vUv = decodeUV(vUv);
-			#endif
-		#endif`,
-
 		ShaderChunk.uv2_vertex,
 		ShaderChunk.color_vertex,
 		ShaderChunk.beginnormal_vertex,
 
-		`#ifdef USE_PACKED_NORMAL
-			objectNormal = decodeNormal(objectNormal);
-		#endif
+		// Shader src: `geometry_decode_normal_vertex`
+		ShaderChunk.geometry_decode_normal_vertex, 
 
-		#ifdef USE_TANGENT
+		`#ifdef USE_TANGENT
 			vec3 objectTangent = vec3( tangent.xyz );
-		#endif
-		`,
+		#endif`,
 
 		ShaderChunk.morphnormal_vertex,
 		ShaderChunk.skinbase_vertex,
@@ -814,11 +747,8 @@ function PackedPhongMaterial(parameters) {
 
 		ShaderChunk.begin_vertex,
 
-		`#ifdef USE_PACKED_POSITION
-			#if USE_PACKED_POSITION == 0
-				transformed = ( vec4(transformed, 1.0) * quantizeMatPos ).xyz;
-			#endif
-		#endif`,
+		// Shader src: `geometry_decode_vertex`
+		ShaderChunk.geometry_decode_vertex, 
 
 		ShaderChunk.morphtarget_vertex,
 		ShaderChunk.skinning_vertex,
@@ -837,6 +767,7 @@ function PackedPhongMaterial(parameters) {
 		"}",
 	].join("\n");
 
+	// Use the original MeshPhongMaterial's fragmentShader. 
 	this.fragmentShader = [
 		"#define PHONG",
