+++ src/renderers/shaders/ShaderChunk/geometry_decode_normal_vertex.glsl.js
@@ -1,0 +1,5 @@
+export default /* glsl */`
+#ifdef USE_PACKED_NORMAL
+	objectNormal = decodeNormal(objectNormal);
+#endif
+`;
